did stuff

+++ UnityAngelHacksGame/Assets/Characters/EnemyData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class EnemyData : ScriptableObject
 {
     public string EnemyName;
-    public int maxHp, maxStam;
-    public int atk, def, vit, EnemyId;
+    public float maxHp, maxStam;
+    public float atk, def, vit, EnemyId;
     public int characterId;
     public Sprite highRes, front, right, left, back;
     public GameObject battleCanvas;
